Repositories and tests

+++ src/main/java/models/Comment.java
@@ -11,14 +11,14 @@ import javax.persistence.*;
 public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String comment;
     @ManyToOne
     private Twit twit;
     @ManyToOne
     private User user;
 
-    public Comment(Integer id, String comment, Twit twit, User user) {
+    public Comment(Long id, String comment, Twit twit, User user) {
         this.id = id;
         this.comment = comment;
         this.twit = twit;
